big-refactor-function
