+++ CMakeLists.txt
@@ -77,12 +77,38 @@ add_executable(so_long
         src/map_validation_utils.c
         src/error.c
         src/render_map.c
-        src/init_textures.c
+        src/init_all_textures.c
         src/init_textures_utils.c
         src/key_press_processing.c
         src/init_map_objects.c
-        src/render_objects.c
+        src/print_moves.c
+        src/init_render_data_utils.c
+        src/init_map_object_utils.c
+        src/appoint_textures.c
+        src/filling_map.c
+        src/init_textures.c
+        src/move_config_utils.c
+        src/main_loop_x.c
+        src/main_loop_event.c
+        src/main_loop_mod_utils.c
+        src/main_loop_print_list.c
+        src/enemies_mod_utils.c
+        src/main_loop_event_utils.c
         )
+
+add_executable(so_long_bonus
+        bonus/main_bonus.c
+        bonus/print_moves_bonus.c
+        src/parser.c
+        src/map_validation.c
+        src/map_validation_utils.c
+        src/error.c
+        src/render_map.c
+        src/init_all_textures.c
+        src/init_textures_utils.c
+        src/key_press_processing.c
+        src/init_map_objects.c
+        src/init_render_data_utils.c src/init_map_object_utils.c src/appoint_textures.c src/filling_map.c src/init_textures.c src/move_config_utils.c src/main_loop_x.c src/main_loop_event.c src/main_loop_mod_utils.c src/main_loop_print_list.c src/enemies_mod_utils.c src/main_loop_event_utils.c)
 
 target_compile_options(mlx PRIVATE -O2)
 target_compile_options(ft PRIVATE -O2)
@@ -93,3 +119,6 @@ target_compile_options(so_long PRIVATE -O2
 
 target_link_libraries(so_long ft)
 target_link_libraries(so_long mlx "-framework OpenGL" "-framework AppKit")
+
+target_link_libraries(so_long_bonus ft)
+target_link_libraries(so_long_bonus mlx "-framework OpenGL" "-framework AppKit")
