+++ bonus/main_bonus.c
@@ -1,6 +1,36 @@
 #include "../include/so_long.h"
 
-t_render_data *init_render_data(char *map_file)
+void	render_counter(t_image_textures image_textures,
+					   void *mlx, void *mlx_win);
+
+void	put_enemies_on_map(char **map_char)
+{
+	int counter;
+	int i;
+	int j;
+
+	counter = 0;
+	i = 1;
+	j = 1;
+	while (map_char[i])
+	{
+		while (map_char[i][j])
+		{
+			if (map_char[i][j] == '0')
+				counter++;
+			if (counter > ENEMY_DENSITY)
+			{
+				map_char[i][j] = 'B';
+				counter = 0;
+			}
+			j++;
+		}
+		j = 1;
+		i++;
+	}
+}
+
+t_render_data	*init_render_data(char *map_file)
 {
 	t_render_data	*render_data;
 	char			**map_char;
@@ -12,7 +42,7 @@ t_render_data *init_render_data(char *map_file)
 	map_validation((const char **)map_char);
 	find_map_size(&render_data->y_length, &render_data->x_length,
 				  map_char);
-///	put_enemies_on_map(map_char);
+	put_enemies_on_map(map_char);
 	render_data->mlx = mlx_init();
 	render_data->image_textures = init_all_textures(render_data->mlx);
 	init_map_objects(&render_data, map_char);
@@ -21,9 +51,9 @@ t_render_data *init_render_data(char *map_file)
 	filling_map(render_data, map_char);
 	del_map_char(map_char, render_data->y_length);
 	render_data->mlx_win = mlx_new_window(render_data->mlx,
-			render_data->x_length * SCALE,
-			render_data->y_length * SCALE,
-			"so_long");
+										  render_data->x_length * SCALE,
+										  render_data->y_length * SCALE,
+										  "so_long");
 	init_render_counters(render_data);
 	return (render_data);
 }
@@ -38,9 +68,8 @@ int	main(int argc, char **argv)
 	render_data = init_render_data(argv[1]);
 
 	render_map(render_data);
-
-///	render_counter(render_data->image_textures,
-///				   render_data->mlx, render_data->mlx_win);
+	render_counter(render_data->image_textures,
+				   render_data->mlx, render_data->mlx_win);
 
 	mlx_hook(render_data->mlx_win, 17, 0, win_kill, render_data);
 	mlx_hook(render_data->mlx_win, 2, 0, key_press_processing_x, render_data);
