+++ include/so_long.h
@@ -13,22 +13,20 @@
 #  define BUFF_SIZE 42
 # endif
 
-# define ENEMY_DENSITY 20
+# define ENEMY_DENSITY 10
 # define MAP_LAYERS 4
 
 
 # define PLAYER_SHIFT 10
 
 # define SCALE 100
-# define FPI 20
-# define REDRAW_INTERVAL
 
-# define PLAYER_FREEZ 500
-# define ANIMATION_SPEED 1000
+# define PLAYER_FREEZ 250
+# define PLAYER_ANIMATION_STEPS 10
+# define ANIMATION_SPEED 4000
 
-# define PLAYER_SPEED 500
-# define ENEMY_SPEED 4000
-# define ENEMY_FREEZ 100
+# define ENEMY_FREEZ 500
+# define ENEMY_ANIMATION_STEPS 10
 
 
 /// change path
@@ -71,10 +69,12 @@ typedef struct s_image_textures
 	t_list	*player_x[NUM_PLAYER_TEXTURES_PACK];
 	t_list	*enemy_x[NUM_ENEMY_TEXTURES_PACK];
 
-	t_list	*space_x;
 	t_list 	*wall_x;
 	t_list	*exit_x;
+	t_list	*space_x;
 
+	void	*space_h;
+	void	*space_v;
 	void	*counter_head;
 	void	*counter[10];
 }	t_image_textures;
@@ -112,7 +112,7 @@ typedef struct s_render_data
 	unsigned int		player_speed;
 
 	unsigned int		enemy_freez;
-	unsigned int		enemy_speed;
+	unsigned int		enemy_animation_steps;
 
 	unsigned int		collectibles_speed;
 
@@ -147,10 +147,35 @@ char		**map_parser(char *map_file);
 void		map_validation(char const **map);
 void		extract_map_data(t_render_data *render_data);
 
-t_image_textures	init_textures(void	*mlx);
-void 				render_map(t_render_data *render_data);
+t_image_textures	init_all_textures(void	*mlx);
+void 				render_map2(t_render_data *render_data);
+
+
+
+
+void	render_map(t_render_data *render_data);
+int		key_press_processing_x(int keycode, t_render_data *render_data);
+int		main_loop_x(t_render_data *render_data);
+
+
+
+void	find_map_size(unsigned int *map_height, unsigned int *map_width,
+					  char **map);
+void	init_map_objects(t_render_data **render_data, char **map_char);
+t_map_object	****init_map(unsigned int y_length, unsigned int x_length);
+void			filling_map(t_render_data *render_data, char **map_char);
+void	del_map_char(char **map_char, int y_len);
+void	init_render_counters(t_render_data *render_data);
+
 
 int	win_kill(int keycode, t_render_data *render_data);
-int key_press_processing(int keycode, t_render_data *render_data);
+
+t_coordinates	x_get_target_pos(t_coordinates current_map_pos, int move_dir);
+void	config_shift_texture(t_map_object **map_obj, int move_dir);
+
+void	print_moves(t_render_data *render_data);
+
+
+
 
 #endif
