+++ maps/map_1.ber
@@ -1,5 +1,5 @@
 1111111111111
 10010000000C1
-10P0011111001
+1000011111001
 1P0011E00C001
 1111111111111
